Optimized deleting events
tableView.reloadData() was called before tableView.deleteRowsAtIndexPathWithAnimation, and this was causing an error because the indexPath of the function parameter would be wrong once reloadData() is called, since it reloads the indexing also. The apple docs also say not to call reloadData() while inserting or deleting. I also realized that the events array will not suddenly become unsorted if an event is deleted, so I removed the sortArray() function call and then saved after deleting the row in the tableView.