Added string representations and other useful functions
Added a equality operator for the Event model, in case events will have to be compared. I was developing a feature that would use this check, but decided to scrap that and keep the equality operator. Also made a function to check if two UIImages are equal using their data representations. Finally, overrided the description property of NSObject to provide a description string for any Event in a certain format.