Modified daysLeft() in the model class Event
Added constants daysToStart and daysToEnd to calculate a new daysLeft value. daysToStart (the days until the "event" or "reminder" starts), will be the value displayed if it is greater than 0. Otherwise, we use the number of days left until the end of the  "event" or "reminder". This will make it clearer how far away events are